+++ vitest.config.ts
@@ -14,7 +14,7 @@ export default defineConfig({
       // plugin provides; tests get a fixed stand-in so app modules that read the
       // app config (e.g. `@app/webmcp/tools`) are importable outside a build.
       "virtual:undocs/app-config": r("./test/stubs/app-config.ts"),
-      // Same deal for the user-pages vfs, which `router.ts` and `webmcp/tools.ts`
+      // Same deal for the user-pages vfs, which `router.ts` and `webmcp/tools/`
       // (its "is this a real route?" check) both read.
       "virtual:undocs/user-pages": r("./test/stubs/user-pages.ts"),
     },
